Moved LightningBertClass class to lightning module
